# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wending Planner API",
        "description": "API de gestion pour la préparation du mariage.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/timeline-events": {
            "get": {
                "tags": [
                    "TimelineEvents"
                ],
                "summary": "Liste des événements du planning",
                "operationId": "34006a9bb3bfd9a6e1218f67cf8324b8",
                "responses": {
                    "200": {
                        "description": "Liste des événements, par ordre chronologique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TimelineEvent"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TimelineEvents"
                ],
                "summary": "Créer un événement",
                "operationId": "056ec664794b816d1f71069cdd166d58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TimelineEventInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Événement créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimelineEvent"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/timeline-events/{timeline_event}": {
            "get": {
                "tags": [
                    "TimelineEvents"
                ],
                "summary": "Détail d'un événement",
                "operationId": "c6aa184b18054746dac2574f65cf38a9",
                "parameters": [
                    {
                        "name": "timeline_event",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Événement trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimelineEvent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Événement introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "TimelineEvents"
                ],
                "summary": "Mettre à jour un événement",
                "operationId": "b2e31e6cb66758939d6b630def9af057",
                "parameters": [
                    {
                        "name": "timeline_event",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TimelineEventInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Événement mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimelineEvent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Événement introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TimelineEvents"
                ],
                "summary": "Supprimer un événement",
                "operationId": "54f2781e5c73b37ddd220766aae521ce",
                "parameters": [
                    {
                        "name": "timeline_event",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Événement supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimelineEvent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Événement introuvable"
                    }
                }
            }
        },
        "/api/venues": {
            "get": {
                "tags": [
                    "Venues"
                ],
                "summary": "Liste des lieux",
                "operationId": "a881d18c3a7e18c81ce59a084693b722",
                "responses": {
                    "200": {
                        "description": "Liste des lieux, par ordre alphabétique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Venue"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Venues"
                ],
                "summary": "Créer un lieu",
                "operationId": "ae68cb3e25e8ba6f89cfc1348dd0dc0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VenueInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lieu créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Venue"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/venues/{venue}": {
            "get": {
                "tags": [
                    "Venues"
                ],
                "summary": "Détail d'un lieu",
                "operationId": "bcba983df0460c53dfb42f86cecbd145",
                "parameters": [
                    {
                        "name": "venue",
                        "in": "path",
                        "description": "ID du lieu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lieu trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Venue"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lieu introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Venues"
                ],
                "summary": "Mettre à jour un lieu",
                "operationId": "250d1147d5d205e4b86d315a783df3cf",
                "parameters": [
                    {
                        "name": "venue",
                        "in": "path",
                        "description": "ID du lieu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VenueInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lieu mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Venue"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lieu introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Venues"
                ],
                "summary": "Supprimer un lieu",
                "operationId": "4f0ae9d6b08485715b2d684dab49db0d",
                "parameters": [
                    {
                        "name": "venue",
                        "in": "path",
                        "description": "ID du lieu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lieu supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Venue"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lieu introuvable"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TimelineEvent": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Cérémonie"
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2027-03-20T14:00:00+01:00"
                    },
                    "note": {
                        "type": "string",
                        "example": "Prévoir les alliances",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TimelineEventInput": {
                "required": [
                    "title",
                    "starts_at"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Cérémonie",
                        "maxLength": 255
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Venue": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Domaine des Roses"
                    },
                    "website": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://domaine-des-roses.fr",
                        "nullable": true
                    },
                    "maps_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://maps.google.com/?q=Domaine+des+Roses",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 3500
                    },
                    "note": {
                        "type": "string",
                        "example": "Beau parc, mais salle un peu petite pour 80 invités.",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "VenueInput": {
                "required": [
                    "name",
                    "price"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Domaine des Roses",
                        "maxLength": 255
                    },
                    "website": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "maps_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}